fix(web): resolve TypeScript module loading and Vite configuration issues
- Fix Vite configuration to properly serve TypeScript files with correct MIME types
- Move main.ts from src/ to public/ directory to resolve module resolution
- Update HTML script reference to use correct file path
- Update import paths in main.ts to use relative paths for component imports
- Add ESBuild target configuration for better TypeScript support
- Fix "Failed to load module script: Expected JavaScript but got HTML" error
- Optimize Vite dependencies for standards-ui package
