+++ apps/web/vite.config.js
@@ -3,19 +3,21 @@ import { resolve } from "path";
 import { defineConfig } from "vite";
 
 export default defineConfig({
-  root: ".",
-  publicDir: "public",
+  root: "public",
   build: {
-    outDir: "dist",
+    outDir: "../dist",
     emptyOutDir: true,
-    rollupOptions: {
-      input: {
-        main: resolve(__dirname, "src/main.ts"),
-      },
-    },
   },
   server: {
     port: 8080,
     host: true,
+  },
+  resolve: {
+    alias: {
+      "@": resolve(__dirname, "src"),
+    },
+  },
+  esbuild: {
+    target: "es2020",
   },
 });
